// display-manager/src/app/common/models/AssetTypes.ts
export enum AssetTypes {
    Default = "Default",

    Racing = "Racing",
    Horses = "Horses",
    GreyHounds = "GreyHounds",
    Virtuals = "Virtuals",
    HorseMeetingResult = "HorseMeetingResult",
    GreyhoundsMeetingResult = "GreyhoundsMeetingResult",
    HorseRacingMultiEvent = "HorseRacingMultiEvent",
    GreyhoundsMultiEvent = "GreyhoundsMultiEvent",
    SportsMultiEvent = "SportsMultiEvent",
    Avr = "Avr",
    Eps = "Eps",
    Nrm = "Nrm",
    NonRunners = "NonRunners",

    Sports = "Sports",

    Marketing = "Marketing",
    Promotion = "Promotion",

    Tv = "Tv",
    Carousel = "Carousel",
    Sky = "Sky",
    ManualCreateTemplate = "ManualCreateTemplate"
}